conference: ensure a proper conference lifetime
